requirements.txt
Flask
Pillow
diffusers
transformers
torch
numpy
werkzeug
